requirements.txt
binance-connector
hopsworks
python-dotenv
pandas
numpy